requirements.txt
asgiref==3.5.2
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyhaikunator==0.0.4
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.2.0
